Move game & curb classes to /lib
